requirements.txt
telethon==1.28.5
twilio==8.2.0
python-dotenv==1.0.0 